refactor: moved all the status & login packets bound to client in new structure
