+++ mojang/minecraft/net/netty/post/_761/client.py
@@ -37,87 +37,6 @@ _Slot = Nested(Slot)
 _BitSets = Prefixed(Array(Long()), VarInt())
 
 
-# Status
-@dataclass
-class StatusResponse(Packet):
-    packet_id = 0
-    response: str = field(metadata={"type": _String})
-
-
-@dataclass
-class PingResponse(Packet):
-    packet_id = 1
-    payload: int = field(metadata={"type": Long()})
-
-
-# Login
-@dataclass
-class Disconnect(Packet):
-    packet_id = 0
-    reason: t.Union[dict, list] = field(metadata={"type": _Chat})
-
-
-@dataclass
-class EncryptionRequest(Packet):
-    packet_id = 1
-
-    server_id: str = field(metadata={"type": _String})
-    public_key: bytes = field(metadata={"type": Prefixed(Bytes(), VarInt())})
-    verify_token: bytes = field(metadata={"type": Prefixed(Bytes(), VarInt())})
-
-
-@dataclass
-class LoginSuccess(Packet):
-    @dataclass
-    class Property:
-        name: str = field(metadata={"type": _String})
-        value: str = field(metadata={"type": _String})
-        is_signed: bool = field(metadata={"type": Bool()})
-        signature: str = field(
-            metadata={
-                "type": Optional(_String),
-                "present": lambda ctx: ctx["is_signed"]["value"],
-            }
-        )
-
-    packet_id = 2
-
-    _uuid: uuid.UUID = field(metadata={"type": UUID()})
-    username: str = field(metadata={"type": _String})
-    len_props: int = field(
-        metadata={
-            "type": VarInt(),
-            "value": lambda ctx: len(ctx["properties"]["value"]),
-        },
-        init=False,
-        default=-1,
-    )
-    properties: t.List[Property] = field(
-        metadata={
-            "type": Array(Nested(Property)),
-            "len": lambda ctx: ctx["len_props"]["value"],
-        }
-    )
-
-
-@dataclass
-class SetCompression(Packet):
-    packet_id = 3
-
-    threshold: int = field(metadata={"type": VarInt()})
-
-
-@dataclass
-class LoginPluginRequest(Packet):
-    packet_id = 4
-
-    message_id: int = field(metadata={"type": VarInt()})
-    channel: str = field(metadata={"type": _Identifier})
-    data: bytes = field(
-        metadata={"type": Bytes(), "len": lambda ctx: ctx["__len"]}
-    )
-
-
 # Play
 @dataclass
 class SpanEntity(Packet):
